# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AgentDef": {
        "description": "One row of `agent_defs` (\u00a74). Frozen: a run holds a snapshot of it.",
        "properties": {
          "allowed_tools": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Allowed Tools",
            "type": "array"
          },
          "auto_approve": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RiskLevel"
            },
            "title": "Auto Approve",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_builtin": {
            "default": false,
            "title": "Is Builtin",
            "type": "boolean"
          },
          "max_steps": {
            "default": 20,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "space_id": {
            "title": "Space Id",
            "type": "string"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "name",
          "role",
          "system_prompt",
          "created_at",
          "updated_at"
        ],
        "title": "AgentDef",
        "type": "object"
      },
      "ApprovalResponse": {
        "description": "One approval, as the dialog renders it.",
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "risk": {
            "title": "Risk",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scope": {
            "default": "call",
            "enum": [
              "call",
              "run"
            ],
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "tool",
          "args",
          "risk",
          "status",
          "created_at"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
...
      "CopyAgentRequest": {
        "description": "Where the copy goes. A copy is a new row with a new id, never a built-in.",
        "properties": {
          "space_id": {
            "title": "Space Id",
            "type": "string"
          }
        },
        "required": [
          "space_id"
        ],
        "title": "CopyAgentRequest",
        "type": "object"
      },
...
      "CreateAgentRequest": {
        "additionalProperties": false,
        "description": "A new definition.\n\nNot :class:`~agentbase.store.agents.AgentDef`: a caller must not mint a built-in.",
        "properties": {
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Tools",
            "type": "array"
          },
          "auto_approve": {
            "items": {
              "type": "string"
            },
            "title": "Auto Approve",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "max_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          },
          "system_prompt": {
            "title": "System Prompt",
            "type": "string"
          }
        },
        "required": [
          "name",
          "role",
          "system_prompt"
        ],
        "title": "CreateAgentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolveApprovalRequest": {
        "additionalProperties": false,
        "description": "A decision on one approval. A misspelled field must not be read as a denial.",
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "scope": {
            "default": "call",
            "enum": [
              "call",
              "run"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "approved"
        ],
        "title": "ResolveApprovalRequest",
        "type": "object"
      },
      "RiskLevel": {
        "description": "How much damage a tool call can do. Policy is a *set* of levels, never a threshold.",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "UpdateAgentRequest": {
        "additionalProperties": false,
        "description": "A partial update. Every field optional; omitted fields are untouched.\n\n``None`` is meaningful for `provider` and `model` (back to inheriting), so\n`model_fields_set` separates \"sent as null\" from \"not sent\".",
        "properties": {
          "allowed_tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Tools"
          },
          "auto_approve": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Approve"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "max_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          }
        },
        "title": "UpdateAgentRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentBase sidecar",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agents": {
      "get": {
        "description": "Every definition, enabled or not: this backs the roster editor.",
        "operationId": "list_agents_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentDef"
                  },
                  "title": "Response List Agents Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agents"
      },
      "post": {
        "operationId": "create_agent_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Agent"
      }
    },
    "/agents/{definition_id}": {
      "delete": {
        "operationId": "delete_agent_agents__definition_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Agent"
      },
      "get": {
        "operationId": "get_agent_agents__definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent"
      },
      "patch": {
        "description": "Apply a partial update. A built-in is editable; only the delete path is guarded.",
        "operationId": "update_agent_agents__definition_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent"
      }
    },
    "/agents/{definition_id}/copy": {
      "post": {
        "description": "A copy of one definition on another space's roster.",
        "operationId": "copy_agent_agents__definition_id__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy Agent"
      }
    },
    "/approvals": {
      "get": {
        "description": "Every approval still waiting for an answer, oldest first.",
        "operationId": "list_approvals_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalResponse"
                  },
                  "title": "Response List Approvals Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Approvals"
      }
    },
    "/approvals/{approval_id}": {
      "get": {
        "operationId": "get_approval_approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Approval"
      },
      "post": {
        "description": "Allow or deny a pending call, and release the agent waiting on it.\n\n404 for an unknown id; 409 for one already settled, so a second window's\nclick reads as \"already answered\" rather than \"gone\".",
        "operationId": "resolve_approval_approvals__approval_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Approval"
      }
    },
...
  }
}